Track on CPU events too
To not count dead backends as still running on the CPU we need to
detect that the backend is dead. Starting from PG17 proc->pid is reset
in ProcKill, before this we can check if the process latch is disowned.
Not nice to be poking around in latch internals like this, but all
alternatives seem to involve scanning bestatus array and correlating
pids.

Also makes sense to exclude ourselves as we will always be on CPU while
looking at wait events.